pyproject.toml
[project]
name = "mock-api"
version = "0.1.0"
description = "CG Mock APIs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "cryptography>=44.0.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.6",
    "greenlet>=3.1.1",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.37",
    "uvicorn[standard]>=0.34.0",
]
